+++ core/realtime-core/src/main/kotlin/org/http4k/testing/PolyHandlerTestClient.kt
@@ -1,0 +1,25 @@
+package org.http4k.testing
+
+import org.http4k.core.PolyHandler
+import org.http4k.core.Request
+
+/**
+ * This class is a test client for a [PolyHandler] which allows the invocation of the underlying handlers.
+ */
+class PolyHandlerTestClient(private val poly: PolyHandler) {
+
+    /**
+     * Invoke the HTTP handler of the [PolyHandler] with the given [Request].
+     */
+    fun http(request: Request) = poly.http?.invoke(request) ?: error("http not implemented")
+
+    /**
+     * Invoke the WebSocket handler of the [PolyHandler] with the given [Request].
+     */
+    fun ws(request: Request) = poly.ws?.testWsClient(request) ?: error("ws not implemented")
+
+    /**
+     * Invoke the Server-Sent Events handler of the [PolyHandler] with the given [Request].
+     */
+    fun sse(request: Request) = poly.sse?.testSseClient(request) ?: error("sse not implemented")
+}